+++ src/com/genir/aitweaks/core/features/shipai/CustomShipAI.kt
@@ -120,7 +120,7 @@ class CustomShipAI(val ship: ShipAPI) : ShipAIPlugin {
     override fun getConfig(): ShipAIConfig = ShipAIConfig()
 
     private fun debug() {
-//        drawCircle(ship.location, ship.collisionRadius / 2f, Color.BLUE)
+        drawCircle(ship.location, ship.collisionRadius / 2f, Color.BLUE)
 //        debugPrint.clear()
 
 //        drawTurnLines(ship)